+++ src/components/mypage/ArticleList/BuyTabArticleList/types.ts
@@ -1,0 +1,8 @@
+import type { TradeBuyActivityType } from '@constants'
+import type { ArticlesElement } from '@types'
+
+export type BuyTabArticleListProps = {
+  activityType: TradeBuyActivityType
+  articles: ArticlesElement[]
+  className?: string
+}
