+++ src/components/mypage/ArticleList/ReviewTabArticleList/types.ts
@@ -1,0 +1,6 @@
+import type { ReviewsElement } from '@types'
+
+export type ReviewTabArticleListProps = {
+  reviews: ReviewsElement[]
+  className?: string
+}
