+++ src/components/mypage/ArticleList/SaleTabArticleList/index.tsx
@@ -2,14 +2,7 @@ import { Divider } from '@offer-ui/react'
 import type { ReactElement } from 'react'
 import { Fragment } from 'react'
 import { SaleTabArticle } from '../../Article'
-import type { ArticlesElement, TradeStatus } from '@types'
-
-export type SaleTabArticleListProps = {
-  articles: ArticlesElement[]
-  className?: string
-  hasToken: boolean
-  onChangeTradeStatus(productId: number, status: TradeStatus): void
-}
+import type { SaleTabArticleListProps } from './types'
 
 export const SaleTabArticleList = ({
   articles,
