+++ src/components/mypage/ArticleList/SaleTabArticleList/types.ts
@@ -1,0 +1,8 @@
+import type { ArticlesElement, TradeStatus } from '@types'
+
+export type SaleTabArticleListProps = {
+  articles: ArticlesElement[]
+  className?: string
+  hasToken: boolean
+  onChangeTradeStatus(productId: number, status: TradeStatus): void
+}
