+++ src/components/mypage/BuyPageContent/styled.ts
@@ -1,0 +1,119 @@
+import styled from '@emotion/styled'
+import type { ColorKeys } from '@offer-ui/react'
+import { Divider } from '@offer-ui/react'
+import { Tabs, Tab } from '@components'
+
+export const StyledContentWrapper = styled.div`
+  ${({ theme }): string => `
+    display: flex;
+    gap: 32px;
+
+    ${theme.mediaQuery.tablet} {
+      position: relative;
+      flex-direction: column;
+      gap: 0;
+    }
+  `}
+`
+export const StyledSearchOptionsWrapper = styled.div`
+  ${({ theme }): string => `
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 20px;
+    border-bottom: 1px solid ${theme.colors.grayScale10};
+
+    ${theme.mediaQuery.tablet} {
+      padding: 16px;
+    }
+  `}
+`
+export const StyledTabsList = styled(Tabs.List)`
+  display: flex;
+  gap: 22px;
+`
+export const StyledTab = styled(Tab)`
+  background: transparent;
+  border: none;
+`
+export const StyledStatusButtonLabel = styled.label`
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+`
+export const StyledCircle = styled.span<{ isCurrent: boolean }>`
+  ${({ theme, isCurrent }): string => `
+    content: '';
+    width: 6px;
+    height: 6px;
+    border-radius: 100%;
+    background-color: ${
+      isCurrent ? theme.colors.brandPrimary : theme.colors.grayScale50
+    };
+  `}
+`
+export const StyledStatusButton = styled.button<{ isCurrent: boolean }>`
+  ${({ theme }): string => `
+    border: none;
+    background: transparent;
+    padding: 0;
+    margin: 0 4px 0 8px;
+    cursor: pointer;
+
+    ${theme.mediaQuery.tablet} {
+      margin: 0 4px;
+    }
+  `}
+`
+export const StyledText = styled.span<{
+  isCurrent?: boolean
+  color?: ColorKeys
+}>`
+  ${({ theme, isCurrent, color }): string => `
+    ${theme.fonts.subtitle01B};
+    color: ${
+      color
+        ? theme.colors[color]
+        : isCurrent
+        ? theme.colors.black
+        : theme.colors.grayScale50
+    };
+
+    ${theme.mediaQuery.tablet} {
+      ${theme.fonts.body02B};
+    }
+  `}
+`
+export const StyledUserProductsWrapper = styled.div`
+  ${({ theme }): string => `
+    width: 892px;
+    height: 855px;
+    border: 1px solid ${theme.colors.grayScale10};
+    border-bottom: none;
+
+    ${theme.mediaQuery.tablet} {
+      width: auto;
+      border: none;
+    }
+  `}
+`
+export const StyledProductListPanels = styled(Tabs.Panels)`
+  ${({ theme }): string => `
+    height: 780px;
+    overflow-y: scroll;
+
+    ${theme.mediaQuery.tablet} {
+      height: auto;
+      overflow-y: unset;
+    }
+  `}
+`
+export const StyledDivider = styled(Divider)`
+  ${({ theme }): string => `
+    display: none;
+
+    ${theme.mediaQuery.tablet} {
+      display: block;
+    }
+  `}
+`
