+++ src/components/mypage/BuyPageContent/types.ts
@@ -1,0 +1,6 @@
+export const sortItems = [
+  {
+    code: 'recently',
+    name: '최신순'
+  }
+]
