+++ src/components/mypage/ProfileBox/types.ts
@@ -1,0 +1,5 @@
+import type { MyProfile } from '@types'
+
+export type ProfileBoxProps = {
+  className?: string
+} & MyProfile
